refactor(step3): Lotto 도메인 분리
1.Lotto - 로또 티켓 도메인
2.LottoMachine - 로또 발급 도메인
