+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -1,0 +1,16 @@
+package lotto.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static lotto.domain.LottoNumber.lottoNumbers;
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class LottoMachineTest {
+
+    @Test
+    @DisplayName("로또 발급 검증")
+    void issue() {
+        assertThat(new LottoMachine(lottoNumbers()).issue(new Payment(10000)).size()).isEqualTo(10);
+    }
+}
